+++ tests/test_accuracy.py
@@ -104,7 +104,7 @@ class TestAccuracy:
         mu = sim_mutations_parameters["rate"]
 
         dts, posteriors = tsdate.date(
-            ts, Ne=Ne, mutation_rate=mu, return_posteriors=True
+            ts, population_size=Ne, mutation_rate=mu, return_posteriors=True
         )
         # make sure we can read node metadata - old tsdate versions didn't set a schema
         if dts.table_metadata_schemas.node.schema is None:
@@ -139,6 +139,38 @@ class TestAccuracy:
         Test that we are in the right theoretical ballpark given known Ne
         """
         ts = tskit.Tree.generate_comb(2).tree_sequence
-        dts = tsdate.date(ts, Ne=Ne, mutation_rate=None)
+        dts = tsdate.date(ts, population_size=Ne, mutation_rate=None)
         # Check the date is within 10% of the expected
         assert 0.9 < dts.node(dts.first().root).time / (2 * Ne) < 1.1
+
+    @pytest.mark.parametrize(
+        "bkwd_rate, trio_tmrca",
+        [  # calculated from simulations
+            (-1.0, 0.76),
+            (-0.9, 0.79),
+            (-0.8, 0.82),
+            (-0.7, 0.85),
+            (-0.6, 0.89),
+            (-0.5, 0.94),
+            (-0.4, 0.99),
+            (-0.3, 1.05),
+            (-0.2, 1.12),
+            (-0.1, 1.21),
+            (0.0, 1.32),
+        ],
+    )
+    def test_piecewise_scaling(self, bkwd_rate, trio_tmrca):
+        """
+        Test that we are in the right theoretical ballpark given known Ne,
+        under exponential growth.
+
+        Check coalescence time of a trio instead of a pair, because of
+        https://github.com/tskit-dev/tsdate/issues/230
+        """
+        time = np.linspace(0, 10, 100)
+        ne = 0.5 * np.exp(bkwd_rate * time)
+        ts = tskit.Tree.generate_comb(3).tree_sequence
+        demo = tsdate.demography.PopulationSizeHistory(ne, time[1:])
+        dts = tsdate.date(ts, population_size=demo, mutation_rate=None)
+        # Check the date is within 10% of the expected
+        assert 0.9 < dts.node(dts.first().root).time / trio_tmrca < 1.1